+++ src/edu/uclouvain/core/nodus/compute/assign/MSAAssignment.java
@@ -80,7 +80,7 @@ public class MSAAssignment extends Assignment {
       return false;
     }
 
-    byte scenario = assignmentParameters.getScenario();
+    int scenario = assignmentParameters.getScenario();
 
     // Read the exclusions
     NodeRulesReader er = new NodeRulesReader(virtualNet, scenario);
